requirements.txt
numpy
torch
PyYAML
gymnasium[box2d]
matplotlib
pybullet
